feat(constraints): multiproof verification

+++ examples/constraints_api/src/types.rs
@@ -137,7 +137,7 @@ pub struct InclusionProofs {
     /// these transactions are the leaves of the transactions tree.
     pub transaction_hashes: Vec<TxHash>,
     /// The generalized indeces of the nodes in the transactions tree.
-    pub generalized_indeces: Vec<u64>,
+    pub generalized_indeces: Vec<usize>,
     /// The proof hashes for the transactions tree.
     pub merkle_hashes: Vec<B256>,
 }
